finagle-mux: Propagates `Ignorable` status over the network
Problem

When a BackupRequestFilter raises on a dispatch that has already been sent over
the wire, mux sends a Tdiscarded message so that the remote peer knows that the
work is no longer needed.  When the server receives the message, it interrupts
the dispatch on its end.  However, it loses the `Ignorable` status when it goes
over the network.

Solution

We should inspect the `why` section of the Tdiscarded to see if it's from
BackupRequestFilter, and if it is, then we should mark it as `Ignorable`.

Result

Stats and failure accrual on receivers don't get polluted by BackupRequestFilter
interruptions from the receiver's callers.

JIRA Issues: CSL-6516

Differential Revision: https://phabricator.twitter.biz/D183456