finatra-http: Fix Exception Mapping with "beforeRouting" Filters
Summary: Problem

A possibile infinite recursion was introduced with a change
to the ExceptionManager unwrapping of failures. The underlying
cause is due to a bug with `beforeRouting = true` Filters throwing
exceptions which trigger the pathological behavior.

The exception handling was predicated on a `RouteInfo` object being
present in the local context -- which will not always be true for a
`beforeRouting = true` Filter resulting in an exception being thrown
in mapping the original exception. Because of the new recursion this
triggers a pathological error which results in infinite recursion in the
`ExceptionManager`.

Solution

Update the `ResponseBuilder` to not require a `RouteInfo` when
classifying an Exception (it will stat details only when the RouteInfo is
actually present).

Change the `ExceptionManager` equality logic to check class
equivalency during recursion in the `toResponse` function.

JIRA Issues: CSL-6629, CSL-6620

Differential Revision: https://phabricator.twitter.biz/D189504