finatra-http: Rename default HTTP port value variable
Problem

It was not originally possible to name this default correctly
as the TwitterServer `AdminHttpServer` defined an incorrectly
named default for the `admin.port` flag. That was renamed
which allows for renaming of the `http.port` default in
`c.t.finatra.http.BaseHttpServer`.

Solution

Rename `defaultFinatraHttpPort` to `defaultHttpPort` which follows
the pattern of "default" + flag name.

TBR=true

Differential Revision: https://phabricator.twitter.biz/D193578