finatra-http: Fix location rewriting in `c.t.f.http.filters.HttpResponseFilter`
Problem

1) Relative paths starting with `http` are not rewritten. Example:
`http-relative-path`.

2) Not all URI schemes are detected. If an app uses a custom
scheme, e.g.,  foo://random-path, it is rewritten when it should not be.

Solution

Use the `java.net.URI` to check if the input location header is a valid URI with a
valid scheme.

Result

Relative paths starting with 'http' will be rewritten correctly and valid URIs
with schemes will not be rewritten.

JIRA Issues: TALON-6729

Differential Revision: https://phabricator.twitter.biz/D191448