inject-core: Ability to use ScalaTest's `whenReady` with Twitter Futures
Problem

ScalaTest provides a `whenReady` function that essentially blocks until a
Future is resolved and then passes the result to a function. The `whenReady`
function takes an abstraction of the Future concept called `FutureConcept.`
There are ]implementations of this for Java and Scala Futures but not for
Twitter Futures.

Solution

Add a new trait which can be mixed in to a ScalaTest test called
`WhenReadyMixin`.  This provides an implicit conversion from a Twitter `Future`
to a `FutureConcept`.

Result

Users have a consistent and easy way to deal with async code in ScalaTests when
using Twitter Futures.

Differential Revision: https://phabricator.twitter.biz/D209215