finatra-thrift: Allow java classes to extend ThriftFilter
Problem

It is not currently possible to extend the `c.t.finatra.thrift.ThriftFilter`
from Java since it is defined as a trait with implemented
methods. This ends up being seen as an interface in Java which is
not allowed to have method implementations.

Solution

Added an abstract class that extends ThriftFilter.

Result

Java classes will now be able to extend
`c.t.finatra.thrift.AbstractThriftFilter`

JIRA Issues: CSL-7047

Differential Revision: https://phabricator.twitter.biz/D221534