finatra-thrift: Remove unecessary param for adding Java Thrift controllers
Problem

The `service: Class[_]` argument in the `c.t.finatra.thrift.routing.ThriftRouter#add`
methods was rendered unnecessary some time ago (when Java exception
mapping was introduced). Worse, in one of the methods this param is shadowed
causing even more confusion.

Solution

Remove the parameter. Instead of deprecating the API we fully remove the parameter
from the method signatures because:
a. it has not been necessary for some time and this does not change functionality and
b. it causes undue confusion

JIRA Issues: CSL-7055

TBR=true

Differential Revision: https://phabricator.twitter.biz/D224336