finatra-http|finatra-thrift: Allow extension of DarkTrafficFilterModule
Problem

It is not currently possible to configure the underlying Finagle client
in either `DarkTrafficFilterModule` in either HTTP or Thrift. This prevents
users from being able to configure Finagle concerns not expressed by
the module's API (like ResponseClassification or other things).

Nor is it possible to implement an `enableSampling` method which uses
anything bound to the object graph as it does not have access to the
Injector.

Solution

Update the `DarkTrafficFilterModule` in both HTTP and Thrift to allow for
specifying further configuration of the underlying Finagle client.

Update the `DrakTrafficFilterModule#enableSampling` method to accept
a `c.t.inject.Injector`.

Result

Users of either module have an avenue to provide further Finagle client
configuration underlying the dark traffic filter.

JIRA Issues: CSL-7075

Differential Revision: https://phabricator.twitter.biz/D225897