finatra: Make StatsFilters ResponseClassification aware
Problem

The HTTP and Thrift StatsFilters record stats about responses
but do not take into account any `ResponseClassifier` when
doing so.

Solution

Update the Filters to use a passed in `ResponseClassifier`.
Specify a server-wide default for server-side response classification
via a `ResponseClassifierModule` which users can override to
provide a custom implementation. This will bind a type to the
object graph which is used in two places: to configure the
underlying Finagle server and for use in the StatsFilters.

JIRA Issues: CSL-7004

Differential Revision: https://phabricator.twitter.biz/D219116