http: Remove need for manual 'apply' on injected filter context
Problem

When using injected filters + nested bodies in the route DSL, one was
required to manually call 'apply' due to the compiler confusing the
overloaded 'filter' methods.

The following code does not work
```
filter[MyFilter] {
  // route def
}
```
however, we're able to do
```
filter(new MyFilter) {
  // route def
}
```

In order to use the former, one would have to do:
```
filter[MyFilter].apply {

}
```

Solution

In order to fix this, another overload for the injected filter DSL type
is added. This allows for more fluent filter definition by allowing the
syntax showed in the first example above, mirroring the capabilities of
non-injected filter syntax and prefix syntax.