finatra-thrift: Fixes and improvements for better Java support
Problem

Filtering within Finatra using a controller that extends generated Java code
does not work properly as the filter are installed too "high" in the stack, that
is they are applied on the resultant `Service[Array[Byte], Array[Byte]]`. This
is generally fine for filters which are side-affecting only but does not work
for filters which need to translate a response, e.g. the `ExceptionMappingFilter`
as the Thrift response has already been written to bytes. Additionally, there
is no `FinatraThriftExceptionMapper` which works over generated Java
exceptions from the `finatra_thrift_exceptions.thrift` Thrift IDL.

Solution

With earlier changes to Scrooge it is now possible install filters of type
`Filter.TypeAgnostic` into the request path of the generated Java Thrift
service. Using this we now can install the Finatra ThriftFilter filter chain
properly to be able to properly mutate responses in exception mapping.

Additionally we provide a `beforeFilter` to continue filtering on the
resultant `Service[Array[Byte], Array[Byte]]` which is useful in the case
of installing a dark traffic filter.

Lastly, we create a `FinatraJavaThriftExceptionMapper` for handling
generated Java exceptions defined in the `finatra_thrift_exceptions.thrift`
Thrift IDL.

Result

Improved usage of Finatra Thrift from Java.

JIRA Issues: CSL-4274

TBR=true

Differential Revision: https://phabricator.twitter.biz/D237483