finatra-thrift: Remove `JavaThriftRouter#beforeFilter`
Problem

The `JavaThriftRouter#beforeFilter` was an API compromise to allow
for a TypeAgnostic filter to be applied to the resultant `Service[-R, +R]`
that is being served. Now that we have moved all filters to TypeAgnostic
Filters it is too difficult an API to understand when to use this method.

Solution

Instead, users should simply apply any `Service[-R, +R]` filter in an override
of `configureService`.

Result

The `JavaThriftRouter` filter API is clearer as all Filter methods act the same
way. Users are still able to intentionally filter the `Service[-R, +R]` by way of
`AbstractThriftServer#configureService`.

Differential Revision: https://phabricator.twitter.biz/D245424