inject-[core|server]: Remove deprecated support for arbitrary field binding in test utilities
Problem

The support for field binding using the Guice `BoundFieldModule` in the
`c.t.inject.IntegrationTestMixin` was deprecated in early 2017 for several reasons:

A. Arbitrary injection (`@Inject`) and binding (`@Bind`) can work well for small sets
of tests but becomes potentially unwieldy for large test suites with many layers of
abstraction and can lead to very brittle tests.

B. It breaks encapsulation. The field binding and injection implicitly uses the underlying
injector (either from the server under test or a TestInjector), however since this is
implicit rather than explicit, the fact that the object graph pertains to a **specific**
server or TestInjector can get lost very quickly in a sufficiently large test suite or when
there are multiple servers under tests leading to a lot of confusion and poor tests.

These issues and others lead to the creation of the more explicit `bind[T]` DSL
(https://twitter.github.io/finatra/user-guide/testing/bind_dsl.html)  to which users
were encouraged to switch when the support for Guice's `BoundFieldModule`
was deprecated.

Lastly, the support for arbitrary field binding and injection references the injector
in the `c.t.injector.IntegrationTestMixin#beforeAll()` method. This makes it such that
non-injectable servers cannot be used as servers under test with the Integration or Feature
test mixins.

Solution

This change removes several deprecated parts of the code base:

A. Remove deprecated support for using `@Bind` or `@Inject` ad-hoc in Finatra
tests. This also removes mock reseting for bound fields and has the side-effect
that the server is not automatically started in most tests because of the inadvertent
call to the underlying injector (which starts the server under test). Thus, some tests
may need to manually start their server under test now.

B. Remove the `c.t.inject.server.FeatureTestMixin#RichFuture` which incorrectly blocks
indefinitely on resolving a Future. Users are encouraged to either await directly in their
tests with a timeout or use `c.t.inject.TestMixin#await` which specifies an overridable
default timeout.

Result

Less technical debt in the project and the Finatra integration and feature testing utilities are
able to properly work with non-injectable servers.

JIRA Issues: CSL-7364, DISC-1898

TBR=true

Differential Revision: https://phabricator.twitter.biz/D250325