inject-server: Add lint rule for setting of deprecated util-logging flags
Problem

Ideally users have no reason to set the flags defined by mixing in
`c.t.inject.server.DeprecatedLogging` which extends the `util-logging`
trait: `c.t.logging.Logging`. These flags are meant to configure
JUL and were present when TwitterServer mixed in the trait. Once TwitterServer
was moved to the slf4j-api, this trait was no longer mixed in and thus
the `DeprecatedLogging` trait was introduced to prevent servers which
were setting the flags from failing to start (the default behavior when an
App is passed a flag value for a flag it does not define). We'd like to
eventually remove the `DeprecatedLogging` trait entirely as the Finatra framework
knows nothing about JUL outside of this trait. Users are encouraged to
configure logging on their chosen implementation or within their server
implementation.

Solution

Add a lint rule to alert if user defined values are passed for the
`c.t.logging.Logging` trait flags.

Result

Even more notice for users potentially setting flags unnecessarily.

Differential Revision: https://phabricator.twitter.biz/D256489