Change Finagle and friends to be compatible with Rsc
Problem

Some Scala code in Finagle and friends is incompatible with Rsc.
While the majority of these incompatibilities can be fixed automatically
with the RscCompat rewrite (these fixes will be submitted in a separate
pull request), there are a few problems that stem from missing
functionality or bugs in Rsc.

Recently, I fixed a significant number of tickets in Rsc, but there
is some stuff left for full compatibility with Finagle and friends.
Unfortunately, all that stuff is pretty challenging, which suggests
that it may be more practical to work around.

Solution

I've manually worked around the following tickets on the Rsc
issue tracker:
  * https://github.com/twitter/rsc/issues/91
  * https://github.com/twitter/rsc/issues/100
  * https://github.com/twitter/rsc/issues/224
  * https://github.com/twitter/rsc/issues/263
  * https://github.com/twitter/rsc/issues/268
  * https://github.com/twitter/rsc/issues/273
  * https://github.com/twitter/rsc/issues/294
  * https://github.com/twitter/rsc/issues/331

Result

All Scala code in Finagle and friends will be compatible with Rsc,
at least to the degree required for our current internal Rsc test
suite to pass.

Differential Revision: https://phabricator.twitter.biz/D264602