finatra/inject-server: EmbeddedTwitterServer that fails to start will now continue to throw startup failure on calls that require a successfully started server
Problem

Extensions of EmbeddedTwitterServer can result in undefined behavior,
if initial startup fails and other methods of the server are still
accessed. Example of this are accessing the `injector` or
`externalThriftHostAndPort` on a server that has failed startup,
which will return values as if the server was started correctly.

Solution

Modify the `start()` method to ensure that a startup failure has
not been seen. Otherwise, if `start()` has failed, it will throw
an IllegalStateException specifying that the server had failed
to start, with the original failure as the cause.

Result

EmbeddedTwitterServer methods that require a succcessful start()
will throw an IllegalStateException with the original start failure
cause, if the initial start() of the server fails.

JIRA Issues: CSL-7548

Differential Revision: https://phabricator.twitter.biz/D265543