finatra-kafka: Implicit implementation of `Flaggable[SeekStrategy]` and `Flaggable[OffsetResetStrategy]`
Problem

When creating a flag for type `c.t.finatra.kafka.domain.SeekStrategy` or
`org.apache.kafka.clients.consumer.OffsetResetStrategy` it is not possible
to be parsed into an instance without a defined `Flaggable` for the type.

Solution

Create an implicit implementation of `Flaggable[SeekStrategy]`
and `Flaggable[OffsetResetStrategy]`

Result

Users can now simply define a flag for a `c.t.finatra.kafka.domain.SeekStrategy` as

```
    private val seekStrategyFlag = flag[SeekStrategy](
      "seek.strategy.flag",
      SeekStrategy.RESUME,
      "This is the seek strategy flag"
    )
```

The flag accepts the values `resume`, `beginning`, `rewind`, and `end`.

or for an `org.apache.kafka.clients.consumer.OffsetResetStrategy`

```
  private val offsetResetStrategyFlag = flag[OffsetResetStrategy](
    "offset.reset.strategy.flag",
    OffsetResetStrategy.LATEST,
    "This is the offset reset strategy flag"
  )
```

The flag accepts the values "latest", "earliest", and "none".

JIRA Issues: DHIS-2963

Differential Revision: https://phabricator.twitter.biz/D271098