finatra-http: Introduce `c.t.finatra.http.HttpServerTrait`
Problem

We'd like to maintain symmetry between how users construct Finatra Thrift and HTTP
servers as well as improve the flexibility of the Finatra HttpServer construction.

Solution

Introduce `c.t.finatra.http.HttpServerTrait` which allows for serving a Finagle
`Service[Request, Response]` directly without needing to configure the Finatra
`HttpRouter` akin to the functionality in `c.t.finatra.thrift.ThriftServerTrait`. As
with Thrift, users would be on the hook for filtering and configuring their implemented
service themselves. But it gives the choice to the user to do it manually or through
the Finatra `HttpRouter` DSL.

Additionally, we remove the internal `BaseHttpServer` which had previously served a
similar purpose but was never exposed as a public API.

Other small imports cleaned up as noticed with compiling changes.

Result

Better parity between building a Finatra server in HTTP and Thrift, reducing cognitive
overhead and more flexibility when creating a Finatra HttpServer to just serve a
Finagle `Service[Request, Response]` directly.

JIRA Issues: CSL-7745

Differential Revision: https://phabricator.twitter.biz/D280896