inject-app|server: Fix Flag ordering from Modules, use AtomicBooleans
Problem

There is a mismatch with how Flags are collected from Modules and then added to an
app or a server. We should be prefering the last flag found in sequence if we are
going to be performing a distinct. Additionally, since Flag has no defined equals or
hashcode, we should be distincting by Flag name as this is the discriminator tracked
by `com.twitter.app.Flags`.

Additionally, we'd like to use `AtomicBooleans` to track atomic boolean state.

Solution

Update the code accordingly.

JIRA Issues: CSL-7829

Differential Revision: https://phabricator.twitter.biz/D306897