inject-request-scope: Filter.TypeAgnostic support and updated documentation
Problem

The Finatra custom guice scope to integrate with Futures using Locals
does not provide a `Filter.TypeAgnostic` version of the scoping Filter
nor is the custom `FinagleRequestScope` written as idiomatically as it
could be.

Solution

Introduce a `TypeAgnosticFinagleRequestScopeFilter` and re-write the
`FinagleRequestScope` to use a `let` function for scoping the seeded
scope. Update Finatra Thrift documentation around filtering.

JIRA Issues: CSL-8043

Differential Revision: https://phabricator.twitter.biz/D310395