finatra-http: Replace 'routeToAdminServer' with 'RouteHint'
Problem

Users currently cannot explicitly control which server a test request is sent to
when a URI is used for both an external server and admin server route. The
`routeToAdminServer` parameter does allow forcing a request to go to the admin
server, however there is no complementary functionality for forcing a request to
the external server. When the same URI is used in both servers, it will always be
seen as an admin route.

Solution

Change `routeToAdminServer` to `RouteHint`, which by default tries to be smart,
but also allows users to explicitly specify which server a test request should
go to.

JIRA Issues: CSL-8024

Differential Revision: https://phabricator.twitter.biz/D313984