inject-server/http/thrift: Allow users to specify a `StatsReceiver` implementation in EmbeddedTwitterServer
Problem

There is currently no way for users to specify a `StatsReceiver` implementation when
constructing an EmbeddedTwitterServer as we always bind an `InMemoryStatsReceiver`
for any injectable underlying server. See: https://github.com/twitter/finatra/issues/496

Solution

Allow users to specify a `StatsReceiver` implementation to use with the underlying
server when the server is injectable. This fixes #496.

JIRA Issues: CSL-8162

Differential Revision: https://phabricator.twitter.biz/D315440