finatra-http: Address HttpWarmup utility
Problem

Warming up your service is a recommended practice. To help with this,
Finatra provides a lifecycle hook and a simple HTTP utility, `HttpWarmup`
which can route requests to user-defined endpoints (routes added to the
`HttpRouter`) during the warmup phase before the server has started and
thus before any ListeningServer has been bound to a port.

The point of warmup is to try to exercise code paths in the service before
accepting traffic to help warm the JVM.

The `c.t.finatra.http.routing.HttpWarmup` utility makes it seem like users
should be able to call TwitterServer HTTP admin interface endpoints. However,
this is not the intended behavior as documented:
http://twitter.github.io/finatra/user-guide/http/warmup.html#c-t-finatra-http-routing-httpwarmup
as routing to the global `HttpMuxer` was intended to route to **user-defined**
admin endpoints.

Regrettably, the logic of calling the global `HttpMuxer` is incorrect since the
`AdminHttpServer` is stateful and the `HttpWarmup` utility needs access to
the `AdminHttpServer`'s underlying `HttpMuxer` to properly route requests,
which is not currently exposed.

Solution

Clean up the logic in the `HttpWarmup` utility to function properly. That is,
to allow routing to **user-defined** routes configured on the Finatra `HttpRouter`.
Routing to endpoints of the TwitterServer HTTP admin interface is not supported
(and was never properly supported).

The `HttpWarmup#send` API changes also to encourage users to expect responses
from the warmup requests. Previously, the `responseCallback` function was a defaulted
param of the main argument list which means it did not have to be provided. It is now
moved to a defaulted arg of a second parameter list which should be invoked when
calling the `#send` function. The `forceRouteToAdminHttpMuxers` param is
renamed to simply `admin` to convey routing to user-defined admin routes more
clearly.

We also move adding of user-defined routes to the TwitterServer HTTP admin
interface to the `postInjectorStartup` lifecycle method to make them available
on the TwitterServer `AdminHttpServer#HttpMuxer` before warmup in the
event the  TwitterServer `AdminHttpServer#HttpMuxer` becomes exposed for
users to be able to route requests to before the ListeningServer is bound.

Result

Users can easily route warmup requests to their routes configured on the
`HttpRouter` with the `HttpWarmup` utility. Note: users are still free to choose
how they want to warm up their service and are not required to use the
`HttpWarmup` utility.

JIRA Issues: CSL-8025

Differential Revision: https://phabricator.twitter.biz/D314152