finatra-http: Add `c.t.inject.server.InMemoryStatsReceiverUtility`
Problem

The printing of metrics from the `EmbeddedTwitterServer` is inconsistent with
the `InMemoryStatsReceiver#print` function and ideally we have one authoritative
way to output metrics. Additionally, all of the functions to get, assert or wait on
metrics from an `InMemoryStatsReceiver` implementation bound to an underlying
server are tied to the `EmbeddedTwitterServer` and not re-usable outside of that
context.

Solution

Introduce the `InMemoryStatsReceiverUtility` to provide metrics utilities over an
`InMemoryStatsReceiver`. Switch to using the `InMemoryStatsReceiver#print` function
for printing metics.

JIRA Issues: CSL-3722

Differential Revision: https://phabricator.twitter.biz/D316806