finatra: Add `com.sun.activation` dependency explicitly
Problem

The `javax.activation` library was deprecated in JDK 9 and is fully
removed in JDK 11. However, Finatra still uses this library for mime
type mapping.

Solution

Add an explicit dependency on the replacement `com.sun.activation`
library. This is a medium-term solution while we plan on how to
evolve the framework away from its dependency on `javax.activation`.

Result

Finatra can be used in JDK 11. This fixes #484.

JIRA Issues: CSL-8161

Differential Revision: https://phabricator.twitter.biz/D328724