scrooge-generator: Create ThriftMethodStats once
Problem

The current implementation of per endpoint stats creates a newly scoped
stats receiver and instance of ThriftMethodStats on each request; it
consequently also creates counters for requests/success/failures on each
request. For some StatsReceivers, this may be expensive, or at least,
not trivial.

Solution

This moves the creation of ThriftMethodStats into perMethodStatsFilter,
where a single instance is (optionally) created for the filter, and then
passed into updated recordRequest/recordResponse methods.

In addition to being more efficient, this also deduplicates code a
little bit.

It's worth noting that when perEndpointStats is false, it appears that
both recordRequest and recordResponse are no-ops; I considered
reorganizing the code to only call them if perEndpointStats is true
(e.g. methodStats.foreach(recordRequest) rather than
recordRequest(methodStats)) but opted to keep the scope of the change
smaller for now.

Signed-off-by: Kevin Oliver <koliver@twitter.com>

JIRA Issues: CSL-8330

Differential Revision: https://phabricator.twitter.biz/D329109