finatra/httpclient: introduce new HttpClientModuleTrait and deprecate HttpClientModule
Problem

The existing `HttpClientModule` provides a default
binding for the `HttpClient` and a `Service[Request, Response]`.
This is not ideal for services which want to bind multiple
HTTP clients. The `HttpClientModule` also does not allow for
any customization or configuration of the underlying
Finagle `Http.Client`.

Solution

With the introduction of the `StackClientModuleTrait` we
have a common base that mimics the existing
`ThriftClientModuleTrait`. Here we create a `HttpClientModuleTrait`
that gives access to configure and customize the underlying
Finagle `Http.Client`, as well as provide custom binding
annotations for `c.t.finagle.Http.Client`,
`c.t.finagle.Service[Request, Response]`, and/or `HttpClient`.

Result
Users will have more control over binding and configuration
of the Finatra `HttpClient`.

JIRA Issues: CSL-7815

Differential Revision: https://phabricator.twitter.biz/D338320