finatra/http: Enrich StreamingResponse and public it
Problem

In an effort to add headers to the StreamingResponse, there is no way to add
headers in controllers without converting it to Future[Response].

Solution

Move the Response metadata to StreamingResponse ctor parameter list.

Result

The new StreamingResponse is in good shape, public it and deprecate the
old one.

JIRA Issues: CSL-7856

Differential Revision: https://phabricator.twitter.biz/D342703