[finatra-jackson] Add optional fields parameter to MethodValidation annotation
Problem

The existing `MethodValidation` annotation doesn't provide a way to expose the fields on which
validation is being performed.

Solution

Added a `fields` parameter to the `MethodValidation` annotation which can be used to specify the
fields on which the validation method is being run. If the validation fails the fields are exposed
out as a part of the Exception that is thrown. By default the value is empty to maintain
compatibility with existing usages.

Results

We have a way to specify fields for which the `MethodValidation` is being performed. This addresses
issue #388.

JIRA Issues: CSL-7494

Differential Revision: https://phabricator.twitter.biz/D338079