inject-app: Update `c.t.inject.app.App` to only recurse through modules once
Problem

An application can be composed of `TwitterModule`s which themselves can specify
other `TwitterModule`s to be included in the application. All of these modules are recursed
through to capture all module-defined flags and to find all modules to use in creation of the Injector.

However, we are incorrectly recursing through the list of modules multiple times. When this happens,
we potentially create new instances of the second-order `TwitterModule`s referenced from any first-order
`TwitterModule`. If any second-order `TwitterModule` defines a flag then original instantiation has it's
flag properly added to the application's Flags map but the second instantiation is the one used to create
the injector and the held flag reference there is never parsed, making it appear that Flags are not working
properly.

This is especially acute in Java. For instance, if you have a Module defined like so:

```
public class TestModuleA extends TwitterModule {

  private Flag<Integer> moduleAFlag = createMandatoryFlag(
      "moduleA.flag",
      "help text",
      "usage text",
      Flaggable.ofJavaInteger()
  );

  @Override
  public void configure() {
    assert moduleAFlag.isDefined();
    assert moduleAFlag.apply() != null;
  }
}
```

And another defined:

```
public class TestModuleB extends TwitterModule {
  private Flag<Float> moduleBFlag = createMandatoryFlag(
      "moduleB.flag",
      "help text",
      "usage text",
      Flaggable.ofJavaFloat()
  );

  @Override
  public Seq<Module> modules() {
    ArrayList<Module> m = new ArrayList<>();
    m.add(new TestModuleA());
    return JavaConverters.asScalaIteratorConverter(m.iterator()).asScala().toSeq();
  }

  @Override
  public void configure() {
    assert moduleBFlag.isDefined();
    assert moduleBFlag.apply() != null;
  }
}
```

If only `TestModuleB` is added to the server,  `TestModuleA` is instantiated twice: once to
find all the module-defined flags and once to finally build a list of all the modules. The second
time the `TestModuleA` is instantiated the Flag from the first reference is the one added to the
application's list of Flags, but the second instantiation is the one used in the application and
thus it's flag reference is never parsed.

Additionally, we don't provide good Java ergonomics for defining a list of second-order `TwitterModule`s
from a `TwitterModule` like we do with servers.

Solution

Update `c.t.inject.app.App` to ensure we recurse through the list of `TwitterModule`s for
an application only once with the results encapsulated into a new object, `Modules`. This instance can
then be used to both find all module-defined flags and to create a the application's `InstalledModules` instance
as  it contains all of the necessary information.

Also, we add `TwitterBaseModule#javaModules` which accepts a list of modules as a Java collection.

Result

We are able to properly account for all modules and flags correctly and better Java support for creation of
`TwitterModule`s which include other `TwitterModule`s.

JIRA Issues: CSL-8587

Differential Revision: https://phabricator.twitter.biz/D349587