inject-server: Update c.t.inject.server.FeatureTestMixin#afterAll
Problem/Solution

We are using application logging to express the exceptions that
occurred on shutdown of the server under test. This is potentially
hidden when tests are not configured for logging. Like the rest
of the testing framework logging we should output this to stderr
such that it is obvious it is from the testing framework rather
than the application stack.

Differential Revision: https://phabricator.twitter.biz/D358588