finatra/inject-logback: Fix NullReferenceException for Gauge
Problem

When logging is misconfigured via an incorrect logback.xml file,
this can cause `LogbackAsyncAppenderBase`'s underlying `blockingQueue` to
not get properly instantiated when the `start` method is called. This leads
to one of the debug gauges ("current_queue_size") throwing a
`NullReferenceException` when the metric is attempted to be collected.

Solution

Guard against the `blockingQueue` being null within the function which
calculates the value of the "current_queue_size" gauge.

Result

There should no longer be a `NullReferenceException` on metrics collection.

JIRA Issues: CSL-8832

Differential Revision: https://phabricator.twitter.biz/D369234