[finatra-kafka] Sanitize topic name in MonitoringConsumer stats scope
Problem:

MonitoringConsumer does not sanitize scope of stats the same way
as the rest of the other stats are generated for Kafka consumers.
Resulting in metrics that are scoped differently within the Kafka consumer
in the application.

Solution:

Use the existing `sanitizeMetricName` method to sanitize
stats scope.

Result:

MonitoringConsumerInterceptor stats are sanitized in a consistent way

Differential Revision: https://phabricator.twitter.biz/D373402