finatra-http: Handle URI decoding issues when extracting path parameters
Problem

It is possible for HTTP routing to return with an Exception before
routing to a user-defined callback. This happens when the `PathPattern`
logic is unable to extract the captured value due to a URI decoding
issue.

A workaround for users is to install the `ExceptionMappingFilter` as
a "beforeRouting" filter which allows them to catch these exceptions
but we should strive to "do the right thing" for users here.

Solution

Handle the case of the `PathPattern` extraction logic throwing an
exception while extracting the capture value. If users want to handle
this case they can by choosing to install the `ExceptionMappingFilter`
as a "beforeRouting" filter but it is not required. The framework
by default should now do the correct thing however, users are free to
override that behavior in this manner.

Result

If the `PathPattern` extraction fails due to a URI decoding error,
the framework will return a `400 - BAD REQUEST` response by default.
This fixes #507.

JIRA Issues: CSL-8765

Differential Revision: https://phabricator.twitter.biz/D381357