finatra: Make the StreamingServerFeatureTest more robust
Problem

We're seeing some timeouts for the StreamingServerFeatureTest in Travis
but not in our local CI which suggests we're seeing some slowness from
the Travis environment.

Solution

Up the timeout for the test. We also should properly segment our
messages and not rely on the HTTP implementation failing to buffer any
messages.

JIRA Issues: CSL-9002

Differential Revision: https://phabricator.twitter.biz/D391713