finatra-http: Deprecate `c.t.finatra.http.modules.DocRootModule`
Problem

The `DocRootModule` defines configuration flags for the
`c.t.finatra.utils.FileResolver` which is used to resolve files either from
the local filesystem or as classpath resources. The `FileResolver` was moved
from `finatra/http` to a more correctly generic location in `finatra/utils`
such that it can be used independent of any HTTP concerns. However,
configuration for injection of a properly configured `FileResolver` is still
incorrectly tied to HTTP because of the location of the `DocRootModule`.

Solution

We are deprecating the `DocRootModule` and introducing the
`c.t.finatra.modules.FileResolverModule` which is defined closer to the
`c.t.finatra.utils.FileResolver` in `finatra/utils`.

Result

This allows for the `FileResolver` to be properly configured and used outside
of any HTTP concerns.

JIRA Issues: CSL-8992

Differential Revision: https://phabricator.twitter.biz/D390932