finatra-http: Don't treat String differently in StreamingResponse
Problem

StreamingResponse used to pass String messages as what they are. If a JSON
format is required, users need to form it from it as JSON.  This special case
leads to some confusion as all other types (expect Buf) are formed into JSON
automatically by `StreamingResponse`.

Solution

Don't treat String differently.

Result

If concatenating a message is the case, convert the message to Buf can do the
work. For example,

```
//controller:
    val reader1 = Reader.fromSeq(Seq("This", "is", "not", "a", "sentence"))
    val streamingResponse1 = responseBuilder.streaming(reader1)

//client:
   response.withBody("""["This","is","not","a","sentence"]""")

```

```
//controller:
    val reader2 = Reader.fromSeq(
        Seq(Buf.Utf8("This "), Buf.Utf8("is "), Buf.Utf8("a "), Buf.Utf8("sentence.")))
    val streamingResponse2 = responseBuilder.streaming(reader2)
//client:
   response.withBody("This is a sentence.")
```

JIRA Issues: CSL-9033

Differential Revision: https://phabricator.twitter.biz/D394904