twitter-server: introduce ability to prevent the `AdminHttpServer` from starting
Problem

The `AdminHttpServer`contains mutable state that allows for
a server to be started again after it is closed. Additionally, there
are lifecycle concerns within Finatra that allow for disabling the
`AdminHttpServer` from starting does so by always starting the
`AdminHttpServer` and closing it post-warmup. Finatra also
suffers from a flaky initialization order issue if the
`disableAdminHttpServer` property is defined as a val.

Solution

We handle the case where an `AdminHttpServer` can be closed,
then re-assigned/re-opened by ensuring that it can only be closed
once and fail traffic after closing. Additionally, we thread the lifecycle
concerns of disabling the `AdminHttpServer` through the
`AdminHttpServer` and ensure that if the server is disabled, it is
never started. We also ensure that `val` and `def` overrides of
`disableAdminHttpServer` behave consistently.

JIRA Issues: CSL-9025

Differential Revision: https://phabricator.twitter.biz/D397925