[logging] lift log4j-over-slf4j from libraries into binaries
Problem

slfj is a logging facade with several implementations - which may be
loaded at runtime. It is generally incorrect to include an
implementation in a library. Instead, include just the API in library
code and allow applications to depend on a specific implementation in the binary

Solution

This diff lifts one of the implementations from libraries into binaries

JIRA Issues: AITOOLS-5215

Differential Revision: https://phabricator.twitter.biz/D402308