[logging] remove some unused dependencies
Problem

slfj is a logging facade with several implementations - which may be loaded at runtime. It is generally incorrect to include an implementation in a library. Instead, include just the API in library code and allow applications to depend on a specific implementation in the binary

Solution

Remove unused specific logging dependencies from BUILD files. While here
remove one more unused dep (guice-testlib)

Result

This is not designed to be a no-op diff. In the case that something
depends on one these two targets, but does not depend on any logging
implementation, things may break.

JIRA Issues: AITOOLS-5215

Differential Revision: https://phabricator.twitter.biz/D406709