inject-app: Add back slf4j-api logging bridges
Problem/Solution

We want to provide 3 slf4j-api logging bridges as
dependencies with inject-app since this an entry
point to creating an App or injectable TwitterServer
using the slf4j-api for logging. We thus explicitly want
the bridges as dependencies here so that users get them
transitively as our recommended logging implementation is
Logback. Let's add them back.

Differential Revision: https://phabricator.twitter.biz/D408547