finatra-thrift: make DarkTrafficModule implementations extend from ThriftClientModuleTrait
Problem

The Thrift DarkTrafficModules are essentially an exposure of a module
to generate a `ThriftMux.Client`. We already have logic for this in the
`inject-thrift-client` project, which exposes a `ThriftClientModuleTrait`.

Solution

Modify the `DarkTrafficFilterModule`s to extend from `ThriftClientModuleTrait`.
In the process, there is some refactoring to keep the functionality of the
`DarkTrafficFilterModule` the same. Additionally, the traits extending from
`StackClientModuleTrait` are modified to allow for better Java support.

Result

Less code duplication and more common configuration APIs across the framework.

JIRA Issues: CSL-9047

Differential Revision: https://phabricator.twitter.biz/D401051