finatra-http: MessageBody API updates
Problem

There are several problems being addressed here which range from
abstract to bugs.

At a high-level, the abstract problem is that the  Finatra HTTP `MessageBodyComponents`
API is not well designed or integrated as it was done after basic HTTP support was
completed, hence HTTP response "rendering" is done in a mix of places including
directly in the `CallbackConverter`, in the `ResponseBuilder`, the default
`MessageBodyWriter` implementation and then any registered customized `MessageBodyWriter`.
The `MessageBodyReader` expresses its API specifically in terms of the finagle-http
`Request` when it should be expressed as the more generic `Message` (in order to also
work for reading `Response` bodies which today has to happen via other mechanisms
involving the `FinatraObjectMapper` directly) and similarly for the `MessageBodyWriter`
which should express its API in terms of a `Message` instead of a finagle-http `Response`.

The problem here lies in that it is not possible for users to swap out all default behavior
as intended through the use of a customized `DefaultMessageBodyWriter` because of the
hardcoded logic in the `CallbackConverter` and the `ResponseBuilder`.

A related issue is that the Finatra Jackson integration is directly handling the marshaling of
message body contents into case classes which is not only one of the sources of an HTTP
dependency in the Finatra Jackson integration, but also means that users are not able to
override this behavior. Most users are using the default behavior hardcoded by the
`FinatraObjectMapper` but this behavior should be overridable through the
`MessageBodyComponents` which is not currently possible as it is hardcoded in
the Finatra Jackson integration and not changeable via Finatra HTTP.

Additionally, we are incorrectly returning a charset encoding parameter on //all// response
content types even though it is only meant to be added to text-based, e.g., `text/*` mime types --
though there is some gray area about `application/json` (https://www.ietf.org/rfc/rfc4627.txt)
and `application/javascript` (JSONP) -- which is generally used instead of its text alternative,
`text/javascript`. Reference: https://tools.ietf.org/html/rfc6657. In short, we should not be
blindly adding the charset encoding parameter which we are currently doing.

Solution

The largest impacting change here is moving the handling of HTTP message body marshaling
from Finatra Jackson to Finatra HTTP. The default behavior which is mostly commonly relied
upon is provided via an implicit which adds methods to the `FinatraObjectMapper` to allow
for usage in the same manner as done today (but with a differently named method which is
necessary due to the inability to overload a method name via an implicit). We have moved
the Finatra HTTP response body handling from the `ResponseBuilder` into the `DefaultMessageBodyWriterImpl`
which means that users can now fully override this behavior.

We also update the `MessageBodyReader` and `MessageBodyWriter` traits to use the more
generic finagle-http `Message` in place of `Request` and `Response` respectively.

Lastly, we fix the charset encoding addendum to response content types to only be applicable
where it should be and not for every content type.

Result

We have not fully addressed response rendering in Finatra HTTP with this change but this
sets the ground work to do so.  Though, users are also able now to completely override the
default Finatra HTTP message body writing which includes all behavior previously hardcoded
in the `ResponseBuilder`. Additionally, we have shrunken the HTTP dependency in Finatra Jackson
and will completely remove it in future changes allowing Finatra Jackson to be usable
with no HTTP dependency.

JIRA Issues: CSL-9138

Differential Revision: https://phabricator.twitter.biz/D400560