[logging] Enable logging implementation checks for logback
Problem

The SLF4J-API is a logging facade with several implementations - which may
be loaded at runtime. It is generally incorrect to include an implementation in
a library. Instead, include just the API in library code and allow applications to
depend on a specific implementation in the binary

Solution

Since many users are not aware of the finer details we have a lint
warning when they try to add a logging implementation to libraries. In
rare cases this is safe or required, so mark said targets as such.

Result

Users will be warned when add a dependency that should be relegated to
runtime only

JIRA Issues: AITOOLS-5215

Differential Revision: https://phabricator.twitter.biz/D413871